pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fusionai"
version = "0.1.0"
description = "FusionAI - AI自动化框架，由湖北福鑫科创信息技术有限公司研发"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "langchain-community>=0.3.19",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.8",
    "langgraph>=0.3.5",
    "readabilipy>=0.3.0",
    "python-dotenv>=1.0.1",
    "socksio>=1.0.0",
    "markdownify>=1.1.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "sse-starlette>=1.6.5",
    "pandas>=2.2.3",
    "numpy>=2.2.3",
    "yfinance>=0.2.54",
    "langchain-deepseek>=0.1.2",
    "langchain-google-genai>=2.0.6",
    "firecrawl-py>=1.6.0",
    "cx_Oracle>=8.3.0",
    "minio>=7.2.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.9",
    "aiofiles>=24.1.0",
    "graphviz>=0.20.3",
    "matplotlib>=3.10.3",
    "networkx>=3.5",
    "plotly>=6.1.2",
    "watchdog>=6.0.0",
    "reportlab>=4.4.1",
    "docx2pdf>=0.1.8",
    "langchain-anthropic>=0.3.15",
]

[project.optional-dependencies]
dev = [
    "black>=24.2.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/build/
'''
